+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Ability to set custom keybindings in config (#135)
 
+### Fixed
+- Don't use cache when URL is typed in bottom bar (#159)
 
 ## [1.7.2] - 2020-12-21
 ### Fixed